requirements.txt
anyio==3.7.0
APScheduler==3.10.4
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
fake-useragent==1.3.0
greenlet==3.0.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
outcome==1.2.0
packaging==23.2
psycopg2-binary==2.9.6
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.17
telegram==0.0.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzlocal==5.1
urllib3==2.0.3
webdriver-manager==4.0.1
wsproto==1.2.0
